refactor code architecture with CSS custom properties, component decomposition and specificity fixes
