+++ src/App.css
@@ -17,7 +17,7 @@
 
 @media (min-width: 768px) {
   .app {
-    max-width: 420px;
+    max-width: var(--app-max-width);
     margin: 0 auto;
     border-radius: 0;
     box-shadow: var(--shadow-lg);
@@ -66,7 +66,7 @@
   background: var(--bg-primary);
   transition: var(--transition-fast);
   color: var(--text-primary);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   box-sizing: border-box;
 }
 
@@ -95,7 +95,7 @@
   transition: var(--transition-fast);
   flex: 1;
   justify-content: center;
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   font-size: var(--text-base);
 }
 
@@ -116,7 +116,7 @@
   border-radius: var(--radius-md);
   cursor: pointer;
   transition: var(--transition-fast);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   font-size: var(--text-base);
   display: flex;
   align-items: center;
@@ -192,7 +192,7 @@
   align-items: center;
   gap: var(--space-xs);
   transition: var(--transition-fast);
-  min-height: 40px;
+  min-height: var(--btn-height-sm);
 }
 
 .reset-current-btn:hover {
@@ -223,7 +223,7 @@
   gap: var(--space-sm);
   transition: var(--transition-fast);
   font-size: var(--text-base);
-  min-height: 48px;
+  min-height: var(--btn-height-lg);
 }
 
 .reset-btn:hover {
