+++ src/styles/components.css
@@ -31,7 +31,7 @@
   color: var(--text-light);
   padding: var(--space-sm) var(--space-md);
   border-radius: var(--radius-lg);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   font-size: var(--text-base);
 }
 
@@ -53,7 +53,7 @@
   color: var(--text-secondary);
   padding: var(--space-sm) var(--space-md);
   border-radius: var(--radius-md);
-  min-height: 40px;
+  min-height: var(--btn-height-sm);
   font-size: var(--text-sm);
 }
 
@@ -67,7 +67,7 @@
   color: var(--text-light);
   padding: var(--space-sm) var(--space-md);
   border-radius: var(--radius-lg);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   font-size: var(--text-base);
 }
 
@@ -82,7 +82,7 @@
   color: var(--text-light);
   padding: var(--space-sm) var(--space-md);
   border-radius: var(--radius-md);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   font-size: var(--text-base);
 }
 
@@ -96,7 +96,7 @@
   color: var(--text-secondary);
   padding: var(--space-sm);
   border-radius: var(--radius-md);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   min-width: 44px;
 }
 
@@ -181,7 +181,7 @@
 .modal {
   background: var(--bg-primary);
   border-radius: var(--radius-2xl);
-  max-width: 400px;
+  max-width: var(--modal-max-width);
   width: 90%;
   max-height: 80vh;
   overflow: hidden;
@@ -203,8 +203,8 @@
 /* Toggle Switch */
 .toggle {
   position: relative;
-  width: 48px;
-  height: 28px;
+  width: var(--toggle-width);
+  height: var(--toggle-height);
   background: var(--bg-gray-dark);
   border-radius: 14px;
   border: none;
@@ -221,8 +221,8 @@
   position: absolute;
   top: 2px;
   left: 2px;
-  width: 24px;
-  height: 24px;
+  width: var(--toggle-slider-size);
+  height: var(--toggle-slider-size);
   background: var(--bg-primary);
   border-radius: var(--radius-full);
   transition: transform 0.3s;
@@ -235,8 +235,8 @@
 
 /* Checkbox */
 .checkbox {
-  width: 28px;
-  height: 28px;
+  width: var(--checkbox-size);
+  height: var(--checkbox-size);
   border: 2px solid var(--bg-gray-dark);
   border-radius: var(--radius-full);
   display: flex;
@@ -245,8 +245,8 @@
   transition: var(--transition-normal);
   background: var(--bg-primary);
   flex-shrink: 0;
-  min-width: 28px;
-  min-height: 28px;
+  min-width: var(--checkbox-size);
+  min-height: var(--checkbox-size);
 }
 
 .checkbox.checked {
@@ -265,7 +265,7 @@
 /* Progress Bar */
 .progress-bar {
   width: 100%;
-  height: 10px;
+  height: var(--progress-bar-height);
   background: rgba(255, 255, 255, 0.2);
   border-radius: 5px;
   overflow: hidden;
@@ -335,7 +335,7 @@
 .list-selector-panel {
   background: var(--bg-primary);
   border-radius: var(--radius-2xl);
-  max-width: 400px;
+  max-width: var(--modal-max-width);
   width: 90%;
   max-height: 80vh;
   overflow: hidden;
@@ -362,7 +362,7 @@
 }
 
 .lists-container {
-  max-height: 300px;
+  max-height: var(--list-max-height);
   overflow-y: auto;
   padding: var(--space-sm);
 }
@@ -699,7 +699,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   min-width: 44px;
 }
 
@@ -747,7 +747,7 @@
   border-radius: var(--radius-md);
   cursor: pointer;
   transition: var(--transition-fast);
-  min-height: 44px;
+  min-height: var(--btn-height-md);
   min-width: 44px;
   display: flex;
   align-items: center;
@@ -774,7 +774,7 @@
   justify-content: center;
   gap: var(--space-sm);
   transition: var(--transition-fast);
-  min-height: 48px;
+  min-height: var(--btn-height-lg);
   font-size: var(--text-base);
 }
 
@@ -802,7 +802,7 @@
   justify-content: center;
   gap: var(--space-sm);
   transition: var(--transition-fast);
-  min-height: 48px;
+  min-height: var(--btn-height-lg);
   font-size: var(--text-base);
 }
 
@@ -866,7 +866,7 @@
   transition: var(--transition-fast);
   flex: 1;
   justify-content: center;
-  min-height: 44px;
+  min-height: var(--btn-height-md);
 }
 
 .export-btn:hover:not(:disabled) {
@@ -900,7 +900,7 @@
   transition: var(--transition-fast);
   width: 100%;
   justify-content: center;
-  min-height: 48px;
+  min-height: var(--btn-height-lg);
 }
 
 .import-btn:hover {
@@ -1013,7 +1013,7 @@
   transition: var(--transition-fast);
   width: 100%;
   justify-content: center;
-  min-height: 48px;
+  min-height: var(--btn-height-lg);
 }
 
 .reset-btn:hover {
@@ -1044,22 +1044,22 @@
 
 /* Mobile sharing styles */
 .share-btn {
-  background: var(--success-green) !important;
-  color: white !important;
+  background: var(--success-green);
+  color: white;
 }
 
 .share-btn:hover {
-  background: var(--success-green-dark) !important;
+  background: var(--success-green-dark);
 }
 
 .action-btn.share-btn {
-  color: var(--success-green) !important;
-  background: none !important;
+  color: var(--success-green);
+  background: none;
 }
 
 .action-btn.share-btn:hover {
-  color: var(--success-green-dark) !important;
-  background: var(--success-bg) !important;
+  color: var(--success-green-dark);
+  background: var(--success-bg);
 }
 
 /* Share Modal Styles */
@@ -1080,7 +1080,7 @@
 .modal-content {
   background: var(--bg-primary);
   border-radius: var(--radius-2xl);
-  max-width: 400px;
+  max-width: var(--modal-max-width);
   width: 90%;
   max-height: 80vh;
   overflow: hidden;
@@ -1176,7 +1176,7 @@
 }
 
 .qr-code-image {
-  max-width: 200px;
+  max-width: var(--qr-code-size);
   width: 100%;
   height: auto;
   border-radius: var(--radius-md);
@@ -1206,13 +1206,13 @@
 }
 
 .backup-btn {
-  background: var(--bg-tertiary) !important;
-  color: var(--text-secondary) !important;
+  background: var(--bg-tertiary);
+  color: var(--text-secondary);
   font-size: 0.875rem;
 }
 
 .backup-btn:hover {
-  background: var(--bg-gray) !important;
+  background: var(--bg-gray);
 }
 
 .qr-modal {
@@ -1233,7 +1233,7 @@
   background: var(--bg-primary);
   border-radius: var(--radius-lg);
   padding: var(--space-xl);
-  max-width: 400px;
+  max-width: var(--modal-max-width);
   width: 100%;
   text-align: center;
   box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1);
@@ -1254,7 +1254,7 @@
 }
 
 .qr-code-image {
-  max-width: 200px;
+  max-width: var(--qr-code-size);
   height: auto;
   border-radius: var(--radius-md);
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
@@ -1321,7 +1321,7 @@
 .confirmation-modal {
   background: var(--bg-primary);
   border-radius: var(--radius-2xl);
-  max-width: 400px;
+  max-width: var(--modal-max-width);
   width: 90%;
   box-shadow: var(--shadow-lg);
   animation: slideUp 0.3s ease-out;
@@ -1422,13 +1422,13 @@
   }
   
   .add-list-btn {
-    min-height: 40px;
+    min-height: var(--btn-height-sm);
     font-size: var(--text-sm);
     padding: var(--space-sm) var(--space-md);
   }
   
   .share-all-btn {
-    min-height: 40px;
+    min-height: var(--btn-height-sm);
     font-size: var(--text-sm);
     padding: var(--space-sm) var(--space-md);
   }
@@ -1439,13 +1439,13 @@
   }
   
   .import-btn {
-    min-height: 40px;
+    min-height: var(--btn-height-sm);
     font-size: var(--text-sm);
     padding: var(--space-sm) var(--space-md);
   }
   
   .reset-btn {
-    min-height: 40px;
+    min-height: var(--btn-height-sm);
     font-size: var(--text-sm);
     padding: var(--space-sm) var(--space-md);
   }
