+++ src/styles/header.css
@@ -12,8 +12,8 @@
   position: absolute;
   top: -50%;
   right: -20%;
-  width: 120px;
-  height: 120px;
+  width: var(--header-decoration-size);
+  height: var(--header-decoration-size);
   background: rgba(255, 255, 255, 0.1);
   border-radius: var(--radius-full);
 }
@@ -48,10 +48,11 @@
   align-items: center;
 }
 
-.settings-btn, .header .share-btn {
-  background: rgba(255, 255, 255, 0.2) !important;
+.header .settings-btn, 
+.header .share-btn {
+  background: rgba(255, 255, 255, 0.2);
   border: none;
-  color: var(--text-light) !important;
+  color: var(--text-light);
   padding: var(--space-sm);
   border-radius: var(--radius-lg);
   cursor: pointer;
@@ -59,18 +60,19 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  min-width: 44px;
-  min-height: 44px;
+  min-width: var(--btn-height-md);
+  min-height: var(--btn-height-md);
 }
 
-.settings-btn:hover, .header .share-btn:hover:not(:disabled) {
-  background: rgba(255, 255, 255, 0.3) !important;
+.header .settings-btn:hover, 
+.header .share-btn:hover:not(:disabled) {
+  background: rgba(255, 255, 255, 0.3);
   transform: scale(1.05);
 }
 
 .header .share-btn:disabled {
-  background: rgba(255, 255, 255, 0.1) !important;
-  color: rgba(255, 255, 255, 0.4) !important;
+  background: rgba(255, 255, 255, 0.1);
+  color: rgba(255, 255, 255, 0.4);
   cursor: not-allowed;
   transform: none;
 }
@@ -90,7 +92,7 @@
   font-size: var(--text-xs);
   margin-top: var(--space-xs);
   max-width: 180px;
-  min-height: 32px;
+  min-height: var(--btn-height-sm);
 }
 
 .list-selector-btn:hover {
